+++ pex/distutils/commands/bdist_pex.py
@@ -105,7 +105,7 @@ class bdist_pex(Command):  # noqa
             target = os.path.join(self.bdist_dir, name + "-" + version + ".pex")
             pex_specs.append((name if name in console_scripts else None, target))
 
-        args = ["-m", "pex", package_dir] + options.requirements + self.pex_args
+        args = ["-m", "pex", "--project", package_dir] + options.requirements + self.pex_args
         if self.get_log_level() < log.INFO and options.verbosity == 0:
             args.append("-v")